Extracts markdown rendering out of ChatMessageBox.tsx
